+++ src/Toml.php
@@ -11,54 +11,86 @@
 
 namespace Yosymfony\Toml;
 
+use Yosymfony\ParserUtils\SyntaxErrorException;
 use Yosymfony\Toml\Exception\ParseException;
 
 /**
- * Parser and dump for Toml format.
+ * Parser for TOML format.
  *
  * @author Victor Puertas <vpgugr@vpgugr.com>
- *
- * @api
  */
 class Toml
 {
     /**
-     * Parse TOML into a PHP array.
+     * Parses TOML into a PHP array.
      *
      * Usage:
      * <code>
-     *  $array = Toml::parse('config.toml');
-     *  print_r($array);
-     *
      *  $array = Toml::parse('key = "[1,2,3]"');
      *  print_r($array);
      * </code>
      *
-     * @return array The TOML converted to a PHP array
+     * @param string $input A string containing TOML
+     *
+     * @return array The TOML converted to a PHP value
+     *
+     * @throws ParseException If the TOML is not valid
      */
-    public static function parse($input)
+    public static function parse(string $input) : array
     {
-        $file = '';
+        $parser = new Parser(new Lexer());
 
-        if (is_file($input)) {
-            if (!is_readable($input)) {
-                throw new ParseException(sprintf('Unable to parse "%s" as the file is not readable.', $input));
+        try {
+            return $parser->parse($input);
+        } catch (SyntaxErrorException $e) {
+            $exception = new ParseException($e->getMessage());
+
+            if ($token = $e->getToken()) {
+                $exception->setParsedLine($token->getLine());
             }
 
-            $file = $input;
-            $input = file_get_contents($input);
+            throw $exception;
+        }
+    }
+
+    /**
+     * Parses a TOML file into a PHP array.
+     *
+     * Usage:
+     * <code>
+     *  $array = Toml::parseFile('config.toml');
+     *  print_r($array);
+     * </code>
+     *
+     * @param string $input A string containing TOML
+     *
+     * @return array The TOML converted to a PHP value
+     *
+     * @throws ParseException If the TOML file is not valid
+     */
+    public static function parseFile(string $filename) : array
+    {
+        if (!is_file($filename)) {
+            throw new ParseException(sprintf('File "%s" does not exist.', $filename));
+        }
+
+        if (!is_readable($filename)) {
+            throw new ParseException(sprintf('File "%s" cannot be read.', $filename));
         }
 
         $parser = new Parser(new Lexer());
 
         try {
-            return $parser->parse($input);
-        } catch (ParseException $e) {
-            if ($file) {
-                $e->setParsedFile($file);
+            return $parser->parse(file_get_contents($filename));
+        } catch (SyntaxErrorException $e) {
+            $exception = new ParseException($e->getMessage());
+            $exception->setParsedFile($filename);
+
+            if ($token = $e->getToken()) {
+                $exception->setParsedLine($token->getLine());
             }
 
-            throw $e;
+            throw $exception;
         }
     }
 }