+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 -----
 * The code has been rewritten from scratch for PHP 7.1.
 * Parse method of `Toml` class returns an empty array in case of empty input.
+* The method `parse` from `Toml` class only can be applied to TOML strings.
+  In case of parsing a TOML filename use the new method `parseFile`.
 * The exceptions have been refactored, so the classes `ExceptionInterface`,
   `LexerException` and `RuntimeException` have been removed.